requirements.txt
streamlit
sentence-transformers
torch
faiss-cpu
pandas
openai
numpy
